Unify both parquet and avro source api
