+++ src/main/scala/com/exasol/cloudetl/scriptclasses/ImportFiles.scala
@@ -5,7 +5,7 @@ import scala.collection.mutable.ListBuffer
 import com.exasol.ExaIterator
 import com.exasol.ExaMetadata
 import com.exasol.cloudetl.bucket._
-import com.exasol.cloudetl.parquet.ParquetSource
+import com.exasol.cloudetl.source._
 
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.hadoop.fs.Path
@@ -22,7 +22,7 @@ object ImportFiles extends LazyLogging {
     val files = groupFiles(iter, 2)
     logger.info(s"Reading file = ${files.take(5).mkString(",")} from bucket = $bucketPath")
 
-    val source = createNewSource(format, files, bucket)
+    val source = createSource(format, files, bucket)
     readAndEmit(source, iter)
   }
 
@@ -37,19 +37,17 @@ object ImportFiles extends LazyLogging {
     files.toSeq
   }
 
-  private[this] def createNewSource(
-    format: String,
-    files: Seq[String],
-    bucket: Bucket
-  ): ParquetSource = {
+  private[this] def createSource(format: String, files: Seq[String], bucket: Bucket): Source = {
     val paths = files.map(f => new Path(f))
     format.toLowerCase match {
+      case "avro"    => AvroSource(paths, bucket.fileSystem, bucket.createConfiguration())
       case "parquet" => ParquetSource(paths, bucket.fileSystem, bucket.createConfiguration())
-      case _         => throw new IllegalArgumentException(s"Unsupported storage format: '$format'")
+      case _ =>
+        throw new IllegalArgumentException(s"Unsupported storage format: '$format'")
     }
   }
 
-  private[this] def readAndEmit(src: ParquetSource, ctx: ExaIterator): Unit =
+  private[this] def readAndEmit(src: Source, ctx: ExaIterator): Unit =
     src.stream.foreach { iter =>
       iter.foreach { row =>
         val columns: Seq[Object] = row.values.map(_.asInstanceOf[AnyRef])
