+++ src/main/scala/com/exasol/cloudetl/source/AvroSource.scala
@@ -1,84 +1,47 @@
 package com.exasol.cloudetl.source
 
+import scala.util.control.NonFatal
+
+import com.exasol.cloudetl.avro.AvroRowIterator
 import com.exasol.cloudetl.data.Row
 
+import com.typesafe.scalalogging.LazyLogging
 import org.apache.avro.file.DataFileReader
 import org.apache.avro.generic.GenericDatumReader
 import org.apache.avro.generic.GenericRecord
-import org.apache.avro.util.Utf8
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.AvroFSInput
 import org.apache.hadoop.fs.FileSystem
 import org.apache.hadoop.fs.Path
 
-final case class AvroSource(paths: Seq[Path], fileSystem: FileSystem, conf: Configuration) {
+/**
+ * An Avro source that can read avro formatted files stored in Hadoop
+ * like distributed storage file systems.
+ */
+final case class AvroSource(
+  override val paths: Seq[Path],
+  override val fileSystem: FileSystem,
+  override val conf: Configuration
+) extends Source
+    with LazyLogging {
 
-  def stream(): Seq[Iterator[Row]] =
+  /** @inheritdoc */
+  override def stream(): Seq[Iterator[Row]] =
     paths.map { path =>
       val avroReader = try {
         createReader(path)
       } catch {
-        case ex: Throwable =>
-          throw new RuntimeException(s"Could not create parquet reader for path $path", ex)
+        case NonFatal(ex) =>
+          logger.error(s"Could not create avro reader for path: $path", ex);
+          throw ex
       }
       AvroRowIterator(avroReader)
     }
 
-  def createReader(path: Path): DataFileReader[GenericRecord] =
+  private[this] def createReader(path: Path): DataFileReader[GenericRecord] =
     new DataFileReader[GenericRecord](
       new AvroFSInput(fileSystem.open(path), fileSystem.getFileStatus(path).getLen),
       new GenericDatumReader[GenericRecord]()
     )
-
-}
-
-object AvroRowIterator {
-
-  def apply(reader: DataFileReader[GenericRecord]): Iterator[Row] =
-    new Iterator[Row] {
-      @SuppressWarnings(Array("org.wartremover.warts.Var"))
-      private[this] var isCompleted = false
-
-      override def hasNext: Boolean =
-        if (isCompleted) {
-          false
-        } else {
-          val hasNext = reader.hasNext
-          if (!hasNext) {
-            reader.close()
-            isCompleted = true
-          }
-          hasNext
-        }
-
-      override def next(): Row = {
-        if (!hasNext) {
-          throw new NoSuchElementException("Avro reader next on empty iterator")
-        }
-        val record = reader.next()
-        recordToRow(record)
-      }
-    }
-
-  private[this] def recordToRow(record: GenericRecord): Row = {
-    val size = record.getSchema.getFields.size
-    val values = Array.ofDim[Any](size)
-    for { index <- 0 until size } {
-      values.update(index, convertRecordValue(record.get(index)))
-    }
-    Row(values.toSeq)
-  }
-
-  @SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
-  private[this] def convertRecordValue(value: Any): Any = value match {
-    case _: GenericRecord =>
-      throw new IllegalArgumentException("Avro nested record type is not supported yet!")
-    case _: java.util.Collection[_] =>
-      throw new IllegalArgumentException("Avro collection type is not supported yet!")
-    case _: java.util.Map[_, _] =>
-      throw new IllegalArgumentException("Avro map type is not supported yet!")
-    case _: Utf8       => value.asInstanceOf[Utf8].toString
-    case primitiveType => primitiveType
-  }
 
 }
