+++ src/main/scala/com/exasol/cloudetl/source/ParquetSource.scala
@@ -1,11 +1,14 @@
-package com.exasol.cloudetl.parquet
+package com.exasol.cloudetl.source
 
 import scala.collection.JavaConverters._
 import scala.language.reflectiveCalls
+import scala.util.control.NonFatal
 
 import com.exasol.cloudetl.data.Row
+import com.exasol.cloudetl.parquet.RowReadSupport
 import com.exasol.cloudetl.util.FileSystemUtil
 
+import com.typesafe.scalalogging.LazyLogging
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.FileSystem
 import org.apache.hadoop.fs.Path
@@ -15,25 +18,41 @@ import org.apache.parquet.hadoop.ParquetReader
 import org.apache.parquet.hadoop.api.ReadSupport
 import org.apache.parquet.schema.MessageType
 
+/**
+ * A companion object to the [[ParquetSource]] class.
+ *
+ * Provides an apply method to create a parquet source.
+ */
 @SuppressWarnings(Array("org.wartremover.warts.Overloading"))
 object ParquetSource {
 
-  def apply(pattern: String, fs: FileSystem, conf: Configuration): ParquetSource =
-    apply(FileSystemUtil.globWithPattern(pattern, fs), fs, conf)
+  def apply(pattern: String, fileSystem: FileSystem, conf: Configuration): ParquetSource =
+    apply(FileSystemUtil.globWithPattern(pattern, fileSystem), fileSystem, conf)
 
 }
 
-final case class ParquetSource(paths: Seq[Path], fs: FileSystem, conf: Configuration) {
+/**
+ * A Parquet source that can read parquet formatted files from Hadoop
+ * compatible storage systems.
+ */
+final case class ParquetSource(
+  override val paths: Seq[Path],
+  override val fileSystem: FileSystem,
+  override val conf: Configuration
+) extends Source
+    with LazyLogging {
 
-  def stream(): Seq[Iterator[Row]] =
+  /** @inheritdoc */
+  override def stream(): Seq[Iterator[Row]] =
     paths.map { path =>
       try {
         using(createReader(path)) { parquetReader =>
           Iterator.continually(parquetReader.read).takeWhile(_ != null)
         }
       } catch {
-        case ex: Throwable =>
-          throw new RuntimeException(s"Could not create parquet reader for path $path", ex)
+        case NonFatal(ex) =>
+          logger.error(s"Could not create parquet reader for path: $path", ex);
+          throw ex
       }
     }
 
@@ -49,17 +68,16 @@ final case class ParquetSource(paths: Seq[Path], fs: FileSystem, conf: Configura
   def getSchema(): Option[MessageType] = {
     val footers = getFooters()
     if (footers.isEmpty) {
-      throw new RuntimeException(
-        s"Could not read parquet metadata at ${paths.take(5).mkString(",")}"
-      )
+      logger.error(s"Could not read parquet metadata from paths: ${paths.take(5).mkString(",")}")
+      throw new RuntimeException("Parquet footers are empty!")
     }
     footers.headOption.map(_.getParquetMetadata().getFileMetaData().getSchema())
   }
 
   def getFooters(): Seq[Footer] =
     paths.flatMap { path =>
-      val status = fs.getFileStatus(path)
-      ParquetFileReader.readAllFootersInParallel(fs.getConf, status).asScala
+      val status = fileSystem.getFileStatus(path)
+      ParquetFileReader.readAllFootersInParallel(fileSystem.getConf, status).asScala
     }
 
   // scalastyle:off structural.type
